Reviewed by Adam.
        Fix build bot.

        * ksvg2/svg/SVGSVGElement.cpp:
        (WebCore::SVGSVGElement::setCurrentScale):



Canonical link: https://commits.webkit.org/18288@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@21930 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ WebCore/ksvg2/svg/SVGSVGElement.cpp
@@ -166,7 +166,7 @@ float SVGSVGElement::currentScale() const
 void SVGSVGElement::setCurrentScale(float scale)
 {
     if (document() && document()->frame())
-        document()->frame()->setZoomFactor(scale / 100.0f);
+        document()->frame()->setZoomFactor(static_cast<int>(scale / 100.0f));
 }
 
 FloatPoint SVGSVGElement::currentTranslate() const
